Minor formatting and diff reduction.
Added a few comments at various places.
Removed 'currentSet' from MLP and instead added a function parameter to
pass the current label to _update_weights.
Simplified _compute_error: It now expects 'target' to be of appropriate
form (this function doesn't seem to be used anywhere though..).
Moved printing of current epoch at the beginning of the training loop.
Fixed: The train function now keeps track of the accuracies regardless
of the verbose parameter. So plotting still works even with
verbose=False.
Clarified the steps of classify.

Tested, and still works ;)